Don't refer as gadget to Inspektor Gadget to avoid confusions
The word "gadget" refers to each of the tools InspektorGadget provides.
Therefore, don't use it to refer to Inspektor Gadget operations to avoid
confusions.

In fact, the next options that will run those tools, will be something
like "Run <gadget-name> gadget" or "Run gadget" which shows the list of
the available gadgets.